refactor(frontend): update nlpEntity formDialog
